store: split out download support to store_download.go
